Refactoring misc classes
