+++ WebApp/src/uk/ac/exeter/QuinCe/data/Instrument/Standards/GasStandardDB.java
@@ -15,12 +15,12 @@ import java.util.Map;
 import javax.sql.DataSource;
 
 import uk.ac.exeter.QuinCe.data.Instrument.RunType;
-import uk.ac.exeter.QuinCe.database.DatabaseException;
-import uk.ac.exeter.QuinCe.database.DatabaseUtils;
-import uk.ac.exeter.QuinCe.database.RecordNotFoundException;
+import uk.ac.exeter.QuinCe.utils.DatabaseException;
+import uk.ac.exeter.QuinCe.utils.DatabaseUtils;
 import uk.ac.exeter.QuinCe.utils.DateTimeUtils;
 import uk.ac.exeter.QuinCe.utils.MissingParam;
 import uk.ac.exeter.QuinCe.utils.MissingParamException;
+import uk.ac.exeter.QuinCe.utils.RecordNotFoundException;
 
 /**
  * Methods for handling records of gas standards in the database.
